Moved common functionality and improved documentation.
# commons/library.py

Moved the functions compress(), place_pairs(), place_indices(), list_labels(), and test_valid_label() to the commons library package.

Improved Observable documentation.

# continuous/library.py

Improved general documentation.
